Implement q_insert_head, q_insert_tail
[What?]
Implement the functionality to add a node to the head
or tail of the queue.

[Why?]
To enable the queue to add nodes with specified strings
to the head or tail.

[How?]
First, check if the queue exists. Then, attempt to
allocate space for a new node, tmp.
If allocation is successful, insert the specified string,
s, into the node.
Finally, use list_add and list_add_tail to add the node to
the queue.