Add allocation for new element's value
[What?]
Allocate space for the value of the new node.

[Why?]
Since the character pointer value in the node also
needs space to store the string s, this functionality is added.

[How?]
Allocate space of size strlen(s) + 1 for the value of the node,
and check if the allocation is successful.
If allocation fails, release the node space and return false.