Add flag on q_delete_dup
[What?]
Add a flag del_flag to delete all adjacent duplicate nodes.

[Why?]
I misunderstood the original requirement, thinking that only one adjacent duplicate node should be kept while deleting the rest.
However, the actual requirement is to delete all adjacent duplicate nodes, so this modification is made.

[How?]
During the iteration of queue nodes, if adjacent duplicate nodes are found,
one of them is deleted, and the del_flag flag is set to true to facilitate the deletion of the last redundant node.