Implement q_reverse
[What?]
Implement a function to reverse the order of all nodes in the queue.

[Why?]
To allow the queue to change direction.

[How?]
Iteratively reverse the direction of all links in the queue.
Also, note that once the pointer l reaches the last node, the program
exits the loop, so an additional action inside the loop should be
added.