Implement q_reverseK
[What?]
Implement a function to reverse the order of every k nodes in the
queue.

[Why?]
To allow the queue to reverse the order of nodes in groups of k.

[How?]
Use a counter to determine if the current pointers l and r are at
the boundary of a group.
If they are, reverse the left group and connect it to the
right group.
Otherwise, simply reverse the pointers.