Revise the q_delete_dup function
[What?]
Modify the q_delete_dup function to fix the error of accessing a null pointer.

[Why?]
Executing list_for_each_entry_safe can traverse the entire list,
but it's essential to note that after traversal, n points to head.
However, head cannot be found using container_of because it's not embedded in the structure.
If container_of is applied to n, it will result in NULL.
Accessing n->value after this will cause an error due to invalid memory access.

[How?]
Add the condition l->next != head to avoid accessing a null pointer.