Implement q_merge
[What?]
Implement a function to merge and sort multiple queues.

[Why?]
This functionality needs to be implemented based on the requirements of the task.

[How?]
In this function, the head passed in points to the first node (queue_contex_t) of this chain.
head->next is first_q, which is used to integrate nodes from other queues.
It is only necessary to traverse all nodes on the entire chain, using list_splice_init to concatenate all queues to first_q.
Finally, the entire queue can be sorted using q_sort.