Implement q_ascend and q_descend
[What?]
Implement a function to remove nodes that do not satisfy the ascending or descending order.

[Why?]
Implement this functionality according to the requirements of the task.

[How?]
q_ascend :
Traverse the entire queue along prev from the tail end of the queue and use curMin to record the smallest string encountered earlier.
If the string of the currently visited node is greater than or equal to curMin, then delete that node.

q_descend :
Traverse the entire queue along prev from the tail end of the queue and use curMax to record the largest string encountered earlier.
If the string of the currently visited node is less than or equal to curMax, then delete that node.