Implement second version of q_swap
[What?]
Propose a second method, and try to utilize the List API as much as possible.

[Why?]
Using the List API can streamline the code and improve code readability.

[How?]
Utilize list_del and list_add_tail to remove the first two nodes from the queue,
and add them back to the queue's tail in reverse order.