Implement approach 2 of the q_delete_mid
[What?]
Propose another version of the method.

[Why?]
The original method relies on the q_size function to obtain
the size of the queue, requiring one traversal of the queue
to determine its size before moving to delete a node in the
middle.
The new method only requires traversing the queue once.

[How?]
Utilize a combination of fast and slow pointers to find the
middle node.
The fast pointer moves two steps at a time while the slow
pointer moves one step at a time.
When the fast pointer finishes traversing the entire queue,
the slow pointer will be positioned at the middle of the queue.