Add the prepare_percentiles function
[What?]
Modify the code for testing program execution time to approximate
the original dudect program.

[Why?]
The original testing program does not implement the
prepare_percentiles function, which may lead to extreme values
affecting the judgment.

[How?]
Add relevant code to fixture.c and make slight modifications to
the function input parameter format.
Since the original code uses the dudect_ctx_t structure to store
information like exec_times, which is not available in lab0-c,
it needs to be passed as a parameter.